requirements.txt
python-telegram-bot==20.7
openai==1.3.7
faiss-cpu==1.7.4
PyPDF2==3.0.1
python-docx==0.8.11
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.0.3